fix code style

+++ tasks/perepelkin_i_convex_hull_graham_scan/omp/include/ops_omp.hpp
@@ -25,7 +25,7 @@ class PerepelkinIConvexHullGrahamScanOMP : public BaseTask {
   static size_t FindPivotParallel(const std::vector<std::pair<double, double>> &pts);
   static void ParallelSort(std::vector<std::pair<double, double>> &data, const std::pair<double, double> &pivot);
 
-  static void DataPartitioning(size_t total_size, const int &threads, std::vector<size_t> &start);
+  static void DataPartitioning(size_t total_size, const int &threads, std::vector<int> &start);
   static void HullConstruction(std::vector<std::pair<double, double>> &hull,
                                const std::vector<std::pair<double, double>> &pts,
                                const std::pair<double, double> &pivot);
